Use rdflib instead of pronto for ontology import
There have been issues importing the official(?) AquaDiva ontology
Pronto throws a lot of warnings during the import, especially
pronto/parsers/rdfxml.py:111: SyntaxWarning:
    unknown element in `owl:ObjectProperty`

The import finally stops with an error
ValueError: invalid datatype:
    'http://www.aquadiva.uni-jena.de/ad-ontology/ad-ontology.0.0/string'

Short of fixing the ontology, the other option was to switch libraries.
The proxy now uses rdflib during the construction of the dictionary tagger,
even though that means to operate on RDF triples instead of ontology concepts.

rdflib itself only throws two warnings, which are due to whitespace in the IRI

http://purl.jp/bio/4/id/200906092801729124#Molecular marker
does not look like a valid URI, trying to serialize this will break.

http://purl.jp/bio/4/id/201206044053997618#ammonia oxidizing archaea
does not look like a valid URI, trying to serialize this will break.

As a side effect, this makes the proxy compatible with Python 3.9.